% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{Sample Metadata}
\format{
A data frame with 24 rows (each representing a unique sample) and
4 columns (representing different metadata variables).
\describe{
\item{Sample.ID}{Character string. Unique identifier for each sample, also serving as the row names (e.g., "JRG_1", "JJG_2").}
\item{Group}{Factor. Indicates the primary experimental group or treatment for each sample (Levels: JRG, JJG, TZG, PAG).}
\item{Origin}{Character string or factor. Represents the geographical or biological origin of the sample (Levels: JR, JJ, TZ, PA).}
\item{Niche}{Character string or factor. Indicates the specific ecological niche or environment from which the sample was collected (Level: G).}
}
}
\source{
Collected during the experimental design phase of the study.
}
\usage{
metadata
}
\description{
A dataset providing essential metadata for each sample included in the
ASV count table. This information is crucial for grouping samples and
performing biologically relevant comparisons.
}
\details{
This metadata table is designed to be directly joined or used in conjunction
with the \code{ASVs} dataset, using \code{Sample.ID} as the linking key, for various
statistical and visualization purposes.
}
\examples{
\dontrun{
# To view the first few rows of the metadata:
head(metadata)

# To check the structure:
str(metadata)

# Get the 'Group' for all samples
metadata$Group
}
}
\keyword{datasets}
